add deals_report
